finny_core/pyproject.toml
[project]
name = "finny_core"
version = "7.2.0"
description = "Duidingslaag tussen administratie-data (MCP/CSV/PDF) en LLM-antwoord — sandwich-pattern: Python rekent, LLM legt uit"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "pydantic>=2.0",
    "pypdf>=3.0",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
